Scribble/requirements.txt
flask
easyocr
google-generativeai
python-dotenv
reportlab
python-docx
opencv-python
numpy